Fixed TS error with mui button overrided styles
