+++ src/styles/app-theme/app.pallete.ts
@@ -1,0 +1,49 @@
+import { blueGrey, green } from '@mui/material/colors'
+
+const palette = {
+  primary: blueGrey,
+  green,
+  backgroundColor: '#F7F7F7',
+  basic: {
+    black: '#000000',
+    white: '#FFFFFF',
+    grey: '#ECEFF1',
+    yellow: '#FFB000',
+    carmenRed: '#B91F1B',
+    burntOrange: '#F56F36',
+    yellowBrown: '#F5D636',
+    yellowGreen: '#9BC541',
+    fruitSalad: '#4CAF50',
+    orientalHerbs: '#12A03A'
+  },
+  companyBlue: 'rgba(0, 167, 167, 0.2)',
+  error: {
+    50: '#FFEFF2',
+    100: '#FFCFD4',
+    200: '#F19B9A',
+    300: '#E77574',
+    400: '#F25750',
+    500: '#F54636',
+    600: '#E63C35',
+    700: '#D32F2F',
+    800: '#C72A28',
+    900: '#B91F1B'
+  },
+  success: {
+    50: '#EBF3EB',
+    100: '#CCE7CD',
+    200: '#A7D5A8',
+    300: '#82C684',
+    400: '#68BA6A',
+    500: '#4CAF50',
+    600: '#44A148',
+    700: '#388D3C',
+    800: '#2C7C32',
+    900: '#1E5F23'
+  },
+  text: {
+    primary: blueGrey[900]
+  }
+}
+
+export default palette
