+++ src/styles/app-theme/app.svgicon.js
@@ -1,4 +1,4 @@
-import palette from './app.pallete.js'
+import palette from './app.pallete'
 
 export const svgIcon = {
   styleOverrides: {
